refactor(snowflake): Extract and compile generator
